backend/requirements.txt
selenium>=4.28.0
beautifulsoup4>=4.12.3
requests>=2.32.3
fake-useragent>=2.0.3
langchain-groq>=0.2.3
transformers>=4.48.1
langchain-core>=0.3.31
pandas>=2.2.3
openpyxl>=3.1.5
orjson>=3.10.15
fastapi>=0.115.6
uvicorn>=0.34.0
pydantic>=2.10.5
python-dotenv>=1.0.1
